+++ contrib/atomicconv/acnn/core/opt_random_tensorgraph.py
@@ -7,7 +7,7 @@ import sys
 import deepchem as dc
 import numpy as np
 import tensorflow as tf
-from deepchem.models.tensorgraph.models.atomic_conv import atomic_conv_model
+from deepchem.models.atomic_conv import atomic_conv_model
 
 sys.path.append("../../models")
 from deepchem.models.tensorgraph.layers import Layer, Feature, Label, L2Loss, AtomicConvolution, Transpose, Dense
