+++ contrib/atomicconv/acnn/refined/tensor_graph_hyper_param_eval.py
@@ -8,7 +8,7 @@ import deepchem as dc
 import json
 import numpy as np
 import tensorflow as tf
-from deepchem.models.tensorgraph.models.atomic_conv import atomic_conv_model
+from deepchem.models.atomic_conv import atomic_conv_model
 
 sys.path.append("../../models")
 from deepchem.models.tensorgraph.layers import Layer, Feature, Label, L2Loss, AtomicConvolution, Transpose, Dense
