+++ deepchem/feat/graph_features.py
@@ -495,7 +495,7 @@ class AtomicConvFeaturizer(ComplexNeighborListFragmentAtomicCoordinates):
       Number of epochs to train the model for
     """
 
-    self.atomic_conv_model = dc.models.tensorgraph.models.atomic_conv.AtomicConvModel(
+    self.atomic_conv_model = dc.models.atomic_conv.AtomicConvModel(
         frag1_num_atoms=frag1_num_atoms,
         frag2_num_atoms=frag2_num_atoms,
         complex_num_atoms=complex_num_atoms,
@@ -546,8 +546,7 @@ class AtomicConvFeaturizer(ComplexNeighborListFragmentAtomicCoordinates):
     # Add the Atomic Convolution layers to fetches
     layers_to_fetch = list()
     for layer in self.atomic_conv_model.layers.values():
-      if isinstance(layer,
-                    dc.models.tensorgraph.models.atomic_conv.AtomicConvolution):
+      if isinstance(layer, dc.models.atomic_conv.AtomicConvolution):
         layers_to_fetch.append(layer)
 
     # Extract the atomic convolution features
